+++ Essenbee.Z80/IBus.cs
@@ -7,7 +7,7 @@ namespace Essenbee.Z80
         IReadOnlyCollection<byte> RAM { get; }
         byte Read(ushort addr, bool ro = false);
         void Write(ushort addr, byte data);
-        byte ReadPeripheral(byte port);
-        void WritePeripheral(byte port, byte data);
+        byte ReadPeripheral(ushort port);
+        void WritePeripheral(ushort port, byte data);
     }
 }
